feat: refactor transaction options handling and add helper functions for chain type conversion
